feat: add paginated response structure and utility function
feat: implement full bulk seed script for realistic dataset generation

feat: create bulk seed script for goals and tasks

feat: define habit and task capacity ports for dependency injection

feat: implement habit capacity adapter for pending habit duration calculation

feat: implement task capacity adapter for committed task duration calculation

feat: add goal title port for batch-fetching goal titles

feat: implement habit capacity adapter for tasks module

+++ src/modules/goals/presentation/dtos/create-goal.dto.ts
@@ -15,7 +15,7 @@ export const CreateGoalSchema = z.object({
     })
     .transform((v) => new Date(v))
     .optional(),
-  estimatedDuration: z.number().positive().optional(),
+  estimatedDuration: z.number().int().positive().optional(),
   estimatedStartDate: z
     .string()
     .datetime({
